vtex-backend/requirements.txt
fastapi
uvicorn
strawberry-graphql
python-dotenv
requests
sqlalchemy